requirements.txt
APScheduler==3.10.4
blinker==1.7.0
cachelib==0.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.2.0
pyarrow==15.0.0
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
six==1.16.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.0
Werkzeug==3.0.1
XlsxWriter==3.1.9
